Fix Selenium issue with checking answers of Classify question
The count (`$i`) of data[Question][answers][$i] was 0, 2, 4, 6 etc.
but is should be 0, 1, 2, 3, 4, etc. so Selenium can iterate trough the answers.

This is now fixed by removing the right (as in left and right) answer
from the $question array. Because it only uses the left answer and
it will get the right answer according to the left answer.
So it shouldn't iterate again over the right answer.