Convert Session to parameteres
The Session resulted in lots of bugs and disappeared sessions.
This resulted in broken API calls to Laravel.
Related to TCP-389

Now the Session is not used anymore, but JavaScript will communicate the UUIDs
via parameters. The UUIDs are stored in JavaScript in TestTake.activeTest

This has impact on the 'taking a test as a student' flow.

The following functions have changed:
- save/save2019
- hand_in
- take/take2019
- take_answer_overview/take_answer_overview2019
- lost_focus
- screenshot_detected