Add more possible event types
Electron reports way more event types and therefore it should be added to the list.
Also, if no translation is available, we fallback to the 'name' entry and show it anyway.
This way, all events will be shown even if there is no translation available.
(so no code change needed for added event, only for translation).