Change way of handling lostFocus reports
We know keep track of the last time a lostFocus reason was reported
insteaed of 'just lostFocus'. This way, if a students triggers two fraud events
within the timeframe, both will be reported instead of only the first one.
Also, there is now a different text for printscreen.

Related to APP-20