Fix logout timeout and reload window for Electron
The logout timeout should use >= instead of ==, because jumping over it
could result in problems.
Also, the timeout should not close the whole Electron app.
Users are not expecting that and results in reporting a crash to support.

When we call logout, we can now either close the app,
or call a reload. window.location.href does not work for Electron,
so we have to use a native call for it.
This native call will be included in the upcoming version (v3.0.2).

Related to APP-40