Add new Bugsnag class and AppVersionDetector
Bugsnag is now upgraded to include much more data while also preserving
privacy better. Now Bugsnag will always include the filters and reporting
metadata is much easier.

There is a new AppVersionDetector that will handle the detection. This is
used by Bugsnag to report version numbers and other valuable data.

The mail 403 is now disabled because it is annoying and not useful.